+++ main/models/__init__.py
@@ -1,0 +1,5 @@
+from .user import User, SpaceUserManager
+from .space import Space, SpaceManager
+from .supporter_membership import SupporterMembership, SupporterMembershipManager
+from .gocardless_mandate import GocardlessMandate, GocardlessMandateManager
+from .gocardless_payment import GocardlessPayment, GocardlessPaymentManager
